+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - This will show you when you can use a gateway.
     - Show you when you do not have a gateway control shard.
 - Remove realm names from Damage Meter.
+- Add `Format` Dropdown for Amounts for the Damage Meter.
 
 # V1.1
 - Add `Keystone Reroll Reminder` to Quality of Life → Toggles.