+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Show you when you do not have a gateway control shard.
 - Remove realm names from Damage Meter.
 - Add `Format` Dropdown for Amounts for the Damage Meter.
+- `Drilldown` from Damage Meter now shows the Specialization Icon.
 
 # V1.1
 - Add `Keystone Reroll Reminder` to Quality of Life → Toggles.