+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix low value abbreviations.
 - Added `Targeted Spells`.
 - Added the ability to colour cast bars when your interrupt is on cooldown, but the cast is interruptible.
+- Track `Hearty Well Fed` for Food.
 
 # V1.1
 - Add `Keystone Reroll Reminder` to Quality of Life → Toggles.